Cleaned up a little bit of code and changing search functions a little bit

+++ src/engine/perft.rs
@@ -6,7 +6,7 @@ use crate::{board::board::Board, moves::movegenerator::generate_moves};
 
 #[allow(dead_code)]
 /// Counts and times the action of generating moves to a certain depth. Prints this information
-pub fn time_move_generation(board: &Board, depth: i32) {
+pub fn time_move_generation(board: &Board, depth: i8) {
     for i in 1..=depth {
         let start = Instant::now();
         print!("{}", count_moves(i, board));
@@ -16,11 +16,9 @@ pub fn time_move_generation(board: &Board, depth: i32) {
     }
 }
 
-pub fn multi_threaded_perft(board: Board, depth: i32) -> usize {
+pub fn multi_threaded_perft(board: Board, depth: i8) -> usize {
     let total = RwLock::new(0);
-
     let moves = generate_moves(&board);
-
     moves.into_vec().into_par_iter().for_each(|m| {
         let mut new_b = board.to_owned();
         new_b.make_move(&m);
@@ -34,8 +32,7 @@ pub fn multi_threaded_perft(board: Board, depth: i32) -> usize {
     x
 }
 
-/// https://www.chessprogramming.org/Perft
-pub fn perft(board: Board, depth: i32) -> usize {
+pub fn perft(board: Board, depth: i8) -> usize {
     let mut total = 0;
     let moves = generate_moves(&board);
     for m in moves.iter() {
@@ -50,7 +47,7 @@ pub fn perft(board: Board, depth: i32) -> usize {
 }
 
 /// Recursively counts the number of moves down to a certain depth
-fn count_moves(depth: i32, board: &Board) -> usize {
+fn count_moves(depth: i8, board: &Board) -> usize {
     let mut count = 0;
     let moves = generate_moves(board);
     if depth == 1 {
